+++ src/BaseParser.php
@@ -1,0 +1,162 @@
+<?php
+
+declare(strict_types=1);
+
+namespace CodeOwners;
+
+use CodeOwners\Exception\UnableToParseException;
+
+/** @internal */
+abstract class BaseParser implements ParserInterface
+{
+    private ParserOptions $options;
+    public function __construct(?ParserOptions $options = null)
+    {
+        $this->options = $options ?? new ParserOptions();
+    }
+
+    /**
+     * @param string $file
+     * @return Pattern[]
+     * @throws UnableToParseException
+     */
+    #[\Override] public function parseFile(string $file): array
+    {
+        return $this->parseIterable($this->getFileIterable($file), $file);
+    }
+
+    /**
+     * @param string $lines
+     * @param ?string $filename
+     * @return Pattern[]
+     * @throws UnableToParseException
+     */
+    #[\Override] public function parseString(string $lines, ?string $filename = null): array
+    {
+        return $this->parseIterable(explode(PHP_EOL, $lines), $filename);
+    }
+
+    /**
+     * @param iterable<string> $lines
+     * @param ?string $filename
+     * @return Pattern[]
+     */
+    private function parseIterable(iterable $lines, ?string $filename = null): array
+    {
+        $patterns = [];
+        $section = null;
+
+        foreach ($lines as $index => $line) {
+            $line = trim($line);
+
+            if ($this->isSection($line)) {
+                if (!$this->options->supportSectionHeaders) {
+                    continue; // skip
+                }
+
+                $section = $this->parseSection($line);
+                continue;
+            }
+
+            $pattern = $this->parseLine($line, new SourceInfo($filename, $index + 1), $section);
+
+            if ($pattern instanceof Pattern) {
+                $patterns[] = $pattern;
+            }
+        }
+
+        return $patterns;
+    }
+
+    private function parseLine(string $line, SourceInfo $sourceInfo, ?Section $section): ?Pattern
+    {
+        $line = trim($line);
+
+        if (str_starts_with($line, '#')) {
+            // comment
+            return null;
+        }
+
+        if ($line === '') {
+            // empty line
+            return null;
+        }
+
+        $regex = $this->options->requireStrictOwners
+            ? '/^(?P<file_pattern>[^\s]+)\s*(?P<owners>(\s+[^\s]*@[^\s]+)*)\s*( #.*)?$/si'
+            : '/^(?P<file_pattern>[^\s]+)(\s+(?P<owners>[^#]*))?/si';
+
+        if (preg_match($regex, $line, $matches) !== 0) {
+            $owners = preg_split('/\s+/', trim($matches['owners'] ?? ''), -1, PREG_SPLIT_NO_EMPTY);
+            if (!is_array($owners)) {
+                // This should not happen as we have full control over the regular expression. In case `preg_split()`
+                // fails an E_WARNING will be emitted by `preg_split()`, so we're not doing that twice.
+                throw new UnableToParseException('Unable to extract owners from line: ' . $line);
+            }
+
+            if (count($owners) === 0 && $section !== null) {
+                $owners = $section->defaultOwners;
+            }
+
+            if ($this->options->requireOwners && count($owners) === 0) {
+                return null;
+            }
+
+            return new Pattern($matches['file_pattern'], $owners, $sourceInfo, $section);
+        }
+
+        return null;
+    }
+
+    /**
+     * @param string $file
+     * @return iterable<string>
+     */
+    private function getFileIterable(string $file): iterable
+    {
+        if (file_exists($file) === false) {
+            throw new UnableToParseException("File {$file} does not exist");
+        }
+
+        if (is_readable($file) === false) {
+            throw new UnableToParseException("File {$file} is not readable");
+        }
+
+        $handle = fopen($file, 'rb');
+        if (is_resource($handle) === false) {
+            throw new UnableToParseException("Unable to create a reading resource for {$file}");
+        }
+
+        while ($line = fgets($handle)) {
+            yield $line;
+        }
+        fclose($handle);
+    }
+
+    private function isSection(string $line): bool
+    {
+        return str_starts_with($line, '[') || str_starts_with($line, '^[');
+    }
+
+    private function parseSection(string $line): ?Section
+    {
+        if (
+            preg_match(
+                '/^(?P<optional>\^)?\[(?P<name>[^\]]+)\](\[(?P<approvals>\d+)\])?(?P<owners>(\s+[^\s]*@[^\s]+)*)$/',
+                $line,
+                $matches
+            )
+        ) {
+            $owners = preg_split('/\s+/', trim($matches['owners'] ?? ''), -1, PREG_SPLIT_NO_EMPTY);
+            if (!is_array($owners)) {
+                throw new UnableToParseException('Unable to extract owners from line: ' . $line);
+            }
+
+            $optional = isset($matches['optional']) && $matches['optional'] === '^';
+            $approvalsRequired = isset($matches['approvals']) ? (int)$matches['approvals'] : null;
+
+            return new Section($matches['name'], $optional, $approvalsRequired, $owners);
+        }
+        return null;
+    }
+}