+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2026-09-08
+### Added
+- Introduce `\CodeOwners\GitLabParser` and `\CodeOwners\GitHubParser` - [#41](https://github.com/timoschinkel/codeowners/pull/41)
+- Add support for spaces in patterns [#41](https://github.com/timoschinkel/codeowners/pull/41)
+
 ## [3.0.0] - 2026-05-04
 ### Changed
 - Marked `\CodeOwners\Exception\NoMatchFoundException` and `\CodeOwners\Exception\UnableToParseException` as final - [#38](https://github.com/timoschinkel/codeowners/pull/38)